+++ app/resources/numismatics/coins/orangelight_coin_builder.rb
@@ -61,7 +61,10 @@ class OrangelightCoinBuilder
         technique_s: decorator.technique,
         weight_s: decorator.weight,
         holdings_1display: holdings_hash,
-        donor_s: decorator.donor
+        donor_s: decorator.donor,
+        numismatic_collection_s: decorator.numismatic_collection,
+        numismatic_accession_s: [decorator.decorated_numismatic_accession&.indexed_label],
+        numismatic_provenance_sm: decorator.provenance
       }
     end
 
@@ -110,7 +113,9 @@ class OrangelightCoinBuilder
         issue_artists_s: parent.artists,
         issue_artists_sort: parent.artists&.first,
         issue_monogram_title_s: parent.decorated_numismatic_monograms.map(&:title),
-        issue_monogram_1display: parent.decorated_numismatic_monograms.map { |m| monogram_hash(m) }.to_json.to_s
+        issue_monogram_1display: parent.decorated_numismatic_monograms.map { |m| monogram_hash(m) }.to_json.to_s,
+        issue_subjects_sm: parent.subjects,
+        issue_date_s: [parent.date_range]
       }
     end